Split Android local AI packages
