+++ apps/android/app/src/main/java/com/flashcardsopensourceapp/app/di/AppGraph.kt
@@ -28,12 +28,12 @@ import com.flashcardsopensourceapp.app.notifications.ReviewNotificationsManager
 import com.flashcardsopensourceapp.app.notifications.AndroidStrictRemindersScheduler
 import com.flashcardsopensourceapp.app.notifications.StrictRemindersManager
 import com.flashcardsopensourceapp.data.local.bootstrap.ensureLocalWorkspaceShell
-import com.flashcardsopensourceapp.data.local.ai.AiChatLiveRemoteService
-import com.flashcardsopensourceapp.data.local.ai.AiChatHistoryStore
-import com.flashcardsopensourceapp.data.local.ai.AiChatPreferencesStore
-import com.flashcardsopensourceapp.data.local.ai.AiCoroutineDispatchers
-import com.flashcardsopensourceapp.data.local.ai.AiChatRemoteService
-import com.flashcardsopensourceapp.data.local.ai.GuestAiSessionStore
+import com.flashcardsopensourceapp.data.local.ai.remote.AiChatLiveRemoteService
+import com.flashcardsopensourceapp.data.local.ai.store.AiChatHistoryStore
+import com.flashcardsopensourceapp.data.local.ai.store.AiChatPreferencesStore
+import com.flashcardsopensourceapp.data.local.ai.remote.AiCoroutineDispatchers
+import com.flashcardsopensourceapp.data.local.ai.remote.AiChatRemoteService
+import com.flashcardsopensourceapp.data.local.ai.store.GuestAiSessionStore
 import com.flashcardsopensourceapp.data.local.cloud.CloudPreferencesStore
 import com.flashcardsopensourceapp.data.local.cloud.remote.CloudRemoteService
 import com.flashcardsopensourceapp.data.local.cloud.sync.SyncLocalStore
