+++ apps/android/data/local/src/main/java/com/flashcardsopensourceapp/data/local/ai/remote/AiCoroutineDispatchers.kt
@@ -1,4 +1,4 @@
-package com.flashcardsopensourceapp.data.local.ai
+package com.flashcardsopensourceapp.data.local.ai.remote
 
 import kotlinx.coroutines.CoroutineDispatcher
