+++ apps/android/data/local/src/main/java/com/flashcardsopensourceapp/data/local/ai/remote/GuestCloudSessionCreator.kt
@@ -1,4 +1,4 @@
-package com.flashcardsopensourceapp.data.local.ai
+package com.flashcardsopensourceapp.data.local.ai.remote
 
 import com.flashcardsopensourceapp.data.local.model.cloud.CloudServiceConfigurationMode
 import com.flashcardsopensourceapp.data.local.model.ai.StoredGuestAiSession
