+++ apps/android/data/local/src/main/java/com/flashcardsopensourceapp/data/local/ai/store/AiChatPreferencesStore.kt
@@ -1,4 +1,4 @@
-package com.flashcardsopensourceapp.data.local.ai
+package com.flashcardsopensourceapp.data.local.ai.store
 
 import android.content.Context
 import androidx.core.content.edit
