+++ apps/android/data/local/src/main/java/com/flashcardsopensourceapp/data/local/ai/wire/AiChatLiveEventRemoteWire.kt
@@ -1,4 +1,4 @@
-package com.flashcardsopensourceapp.data.local.ai
+package com.flashcardsopensourceapp.data.local.ai.wire
 
 import com.flashcardsopensourceapp.data.local.model.ai.AiChatLiveEvent
 import com.flashcardsopensourceapp.data.local.model.ai.AiChatLiveEventMetadata
