+++ apps/android/data/local/src/main/java/com/flashcardsopensourceapp/data/local/ai/wire/AiChatRemoteWireSupport.kt
@@ -1,4 +1,4 @@
-package com.flashcardsopensourceapp.data.local.ai
+package com.flashcardsopensourceapp.data.local.ai.wire
 
 import com.flashcardsopensourceapp.data.local.cloud.wire.CloudContractMismatchException
 import com.flashcardsopensourceapp.data.local.cloud.wire.buildRemoteContractMismatch
