+++ apps/android/data/local/src/main/java/com/flashcardsopensourceapp/data/local/repository/cloudsync/guest/CloudGuestSessionCoordinator.kt
@@ -1,7 +1,7 @@
 package com.flashcardsopensourceapp.data.local.repository.cloudsync.guest
 
-import com.flashcardsopensourceapp.data.local.ai.GuestAiSessionStore
-import com.flashcardsopensourceapp.data.local.ai.GuestCloudSessionCreator
+import com.flashcardsopensourceapp.data.local.ai.store.GuestAiSessionStore
+import com.flashcardsopensourceapp.data.local.ai.remote.GuestCloudSessionCreator
 import com.flashcardsopensourceapp.data.local.bootstrap.ensureLocalWorkspaceShell
 import com.flashcardsopensourceapp.data.local.cloud.CloudPreferencesStore
 import com.flashcardsopensourceapp.data.local.cloud.remote.CloudRemoteException
