+++ apps/android/data/local/src/main/java/com/flashcardsopensourceapp/data/local/repository/cloudsync/progress/CloudProgressRemoteReader.kt
@@ -1,6 +1,6 @@
 package com.flashcardsopensourceapp.data.local.repository.cloudsync.progress
 
-import com.flashcardsopensourceapp.data.local.ai.GuestAiSessionStore
+import com.flashcardsopensourceapp.data.local.ai.store.GuestAiSessionStore
 import com.flashcardsopensourceapp.data.local.cloud.CloudPreferencesStore
 import com.flashcardsopensourceapp.data.local.cloud.remote.CloudRemoteGateway
 import com.flashcardsopensourceapp.data.local.model.cloud.CloudAccountState
