+++ apps/android/data/local/src/main/java/com/flashcardsopensourceapp/data/local/repository/cloudsync/workspace/CloudWorkspaceLinkCoordinator.kt
@@ -1,6 +1,6 @@
 package com.flashcardsopensourceapp.data.local.repository.cloudsync.workspace
 
-import com.flashcardsopensourceapp.data.local.ai.GuestAiSessionStore
+import com.flashcardsopensourceapp.data.local.ai.store.GuestAiSessionStore
 import com.flashcardsopensourceapp.data.local.cloud.CloudPreferencesStore
 import com.flashcardsopensourceapp.data.local.cloud.PendingGuestUpgradeState
 import com.flashcardsopensourceapp.data.local.cloud.remote.CloudRemoteGateway
