+++ apps/android/feature/ai/src/test/java/com/flashcardsopensourceapp/feature/ai/runtime/session/AiChatRuntimeSendDraftRestorationTest.kt
@@ -5,7 +5,7 @@ import com.flashcardsopensourceapp.feature.ai.runtime.conversation.AiConversatio
 import com.flashcardsopensourceapp.feature.ai.runtime.conversation.makeAssistantStatusMessage
 import com.flashcardsopensourceapp.feature.ai.runtime.conversation.makeUserMessage
 import com.flashcardsopensourceapp.feature.ai.runtime.errors.AiAlertState
-import com.flashcardsopensourceapp.data.local.ai.AiChatRemoteException
+import com.flashcardsopensourceapp.data.local.ai.remote.AiChatRemoteException
 import com.flashcardsopensourceapp.data.local.model.ai.AiChatAttachment
 import com.flashcardsopensourceapp.data.local.model.ai.AiChatContentPart
 import com.flashcardsopensourceapp.data.local.model.cloud.CloudAccountState
